[refactor] put all messenger logic into new messenger file and renamed other files to rollup logic names
