+++ messenger/rollup_app.go
@@ -39,12 +39,6 @@ type Config struct {
 	SeqPrivate   string `env:"SEQUENCER_PRIVATE, required"`
 }
 
-// Transaction represents a transaction in the blockchain.
-type Transaction struct {
-	Sender  string `json:"sender"`
-	Message string `json:"message"`
-}
-
 // App is the main application struct, containing all the necessary components.
 type App struct {
 	executionRPC    string
@@ -58,38 +52,6 @@ type App struct {
 	newBlockChan    chan Block
 	wsClients       WSClientList
 	sync.RWMutex
-}
-
-func encodeTx(tx Transaction) ([]byte, error) {
-	// encode transaction to send to sequencer
-	data, err := json.Marshal(tx)
-	if err != nil {
-		log.Errorf("error encoding transaction: %s\n", err)
-		return data, err
-	}
-	return data, nil
-}
-
-func decodeTx(txEncoded []byte) (Transaction, error) {
-	tx := &Transaction{}
-	if err := json.Unmarshal(txEncoded, tx); err != nil {
-		log.Errorf("error decoding transaction: %s\n", err)
-		return *tx, err
-	}
-	return *tx, nil
-}
-
-func GenesisTransaction() []byte {
-	genesisTx := Transaction{
-		Sender:  "astria",
-		Message: "hello, world!",
-	}
-	encodedTx, err := encodeTx(genesisTx)
-	if err != nil {
-		log.Errorf("error encoding genesis tx: %s\n", err)
-		panic(err)
-	}
-	return encodedTx
 }
 
 func NewApp(cfg Config) *App {
@@ -130,9 +92,8 @@ func (a *App) makeExecutionServer() *ExecutionServiceServerV1Alpha2 {
 // setupRestRoutes sets up the routes for the REST API.
 func (a *App) setupRestRoutes() {
 	a.restRouter.HandleFunc("/block/{height}", a.getBlock).Methods("GET")
-	a.restRouter.HandleFunc("/message", a.postMessage).Methods("POST")
-	a.restRouter.HandleFunc("/recent", a.getRecentMessages).Methods("GET")
 	a.restRouter.HandleFunc("/ws", a.serveWS)
+	registerHandlers(a)
 }
 
 // makeRestServer creates a new HTTP server for the REST API.
@@ -174,68 +135,6 @@ func (a *App) getBlock(w http.ResponseWriter, r *http.Request) {
 	}
 
 	w.Write(blockJson)
-}
-
-func (a *App) getRecentMessages(w http.ResponseWriter, _ *http.Request) {
-	var messages []Transaction
-	for i := len(a.rollupBlocks.Blocks); i > 0; i-- {
-		block := a.rollupBlocks.Blocks[i-1]
-		if len(block.Txs) > 0 {
-			log.Infof("block txs: %v", block.Txs)
-			for _, txRaw := range block.Txs {
-				log.Infof("raw txn: %v", txRaw)
-				tx, err := decodeTx(txRaw)
-				if err != nil {
-					log.Errorf("tried to decode malformed message: %s\n", err)
-					continue
-				}
-				messages = append(messages, tx)
-			}
-			if len(messages) >= 100 {
-				break
-			}
-		}
-	}
-
-	// keep only the most recent 100
-	if len(messages) > 100 {
-		messages = messages[len(messages)-100:]
-	}
-
-	messagesJson, err := json.Marshal(messages)
-	if err != nil {
-		log.Errorf("error marshalling messages: %s\n", err)
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-
-	w.Write(messagesJson)
-}
-
-func (a *App) postMessage(w http.ResponseWriter, r *http.Request) {
-	var tx Transaction
-	// decode transaction to ensure proper format
-	err := json.NewDecoder(r.Body).Decode(&tx)
-	if err != nil {
-		log.Errorf("error decoding transaction: %s\n", err)
-		w.WriteHeader(http.StatusBadRequest)
-		return
-	}
-	log.Infof("'postMessage': tx decoded successfully: %v, ", tx)
-	// recode transaction to send to sequencer
-	txEncoded, err := encodeTx(tx)
-	if err != nil {
-		log.Errorf("error re-encoding transaction: %s\n", err)
-		w.WriteHeader(http.StatusBadRequest)
-		return
-	}
-	resp, err := a.sequencerClient.SendMessage(txEncoded)
-	if err != nil {
-		log.Errorf("error sending message: %s\n", err)
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-	log.WithField("responseCode", resp.Code).Debug("transaction submission result")
 }
 
 func (a *App) serveWS(w http.ResponseWriter, r *http.Request) {
@@ -304,34 +203,19 @@ func (a *App) Run() {
 			if len(block.Txs) == 0 {
 				continue
 			}
-			// re-encode messages as json for front end
-			transactions := []Transaction{}
-			for _, encodedTx := range block.Txs {
-				decodedTx, err := decodeTx(encodedTx)
-				if err != nil {
-					log.Errorf("error decoding tx in block: %v, error: %s\n", encodedTx, err)
-					continue
-				}
-				transactions = append(transactions, decodedTx)
-			}
 
-			// only write blocks with valid transactions
-			if len(transactions) == 0 {
+			// decode transactions into format that the client can handle
+			txsJson := prepareBlockForClient(block.Txs)
+
+			if len(block.Txs) == 0 {
+				log.Info("post txs filtering no txs remaining")
 				continue
 			}
-
-			// send transaction encoded format to frontend
-			txsJson, err := json.Marshal(transactions)
-			log.Debugf("marshalled txsJson: %s", txsJson)
-			if err != nil {
-				log.Errorf("Failed to marshal transactions: %v", err)
-			} else {
-				for client := range a.wsClients {
-					select {
-					case client.egress <- txsJson:
-					default:
-						log.Warnf("Could not send transactions to ws client: %s", txsJson)
-					}
+			for client := range a.wsClients {
+				select {
+				case client.egress <- txsJson:
+				default:
+					log.Warnf("Could not send transactions to ws client: %s", txsJson)
 				}
 			}
 		}
